SME/requirements.txt
flask
python-docx
mysql-connector-python
jinja2
python-dotenv
requests
